Rename integration tests
These tests have the same name than other tests in `unit` and
`acceptance` folder.

VS Code IDE in unable to do the discovery process of the tests when:

1. There is different sources of tests (different folders)
2. Some names are duplicated in tests of different folders.

This causes the discovery process to crash, so tests are not shown in
Test Explorer.

Renaming the tests with duplicated names works and let the developers
that uses VS Code to keep working with this IDE.